+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 4.0.0-rc.2
+
+* Add `sendResultError` to `IsolateContactorController` to make it easier to send the Exception from the Isolate to main app.
+* `IsolateException` parameters are now `Object` and `StackTrace`.
+
+## 4.0.0-rc.1
+
+* Add parameter type as `P` and change return type to `R`.
+* Add return type as `void` to `onDispose` method.
+* Change `isolateParams` from dynamic to `Object?`.
+
 ## 3.0.0
 
 * Bumped Dart sdk to `>=2.18.0 <4.0.0`.